Moved java files to src/java/gov/usgs/earthquake/catalog
